requirements.txt
python-binance
configparser
PySimpleGUI